introduced ListComponentFactory.java and moved its dependencies to component package.
