+++ src/org/intellij/ideaplugins/tabswitch/component/ListCellRendererWithColorFactory.java
@@ -1,4 +1,4 @@
-package org.intellij.ideaplugins.tabswitch;
+package org.intellij.ideaplugins.tabswitch.component;
 
 import java.awt.Font;
